iOS SingleLayerComposeScene with feature flag
