+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/ComposeUIViewController.uikit.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.ui.window
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.uikit.ComposeUIViewControllerConfiguration
+import platform.UIKit.UIViewController
+
+fun ComposeUIViewController(content: @Composable () -> Unit): UIViewController =
+    ComposeUIViewController(configure = {}, content = content)
+
+fun ComposeUIViewController(
+    configure: ComposeUIViewControllerConfiguration.() -> Unit = {},
+    content: @Composable () -> Unit
+): UIViewController = ComposeContainer(
+    configuration = ComposeUIViewControllerConfiguration().apply(configure),
+    content = content,
+)
