+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/DensityProvider.uikit.kt
@@ -1,0 +1,83 @@
+/*
+ * Copyright 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.ui.window
+
+import androidx.compose.ui.unit.Density
+import platform.UIKit.UIContentSizeCategoryAccessibilityExtraExtraExtraLarge
+import platform.UIKit.UIContentSizeCategoryAccessibilityExtraExtraLarge
+import platform.UIKit.UIContentSizeCategoryAccessibilityExtraLarge
+import platform.UIKit.UIContentSizeCategoryAccessibilityLarge
+import platform.UIKit.UIContentSizeCategoryAccessibilityMedium
+import platform.UIKit.UIContentSizeCategoryExtraExtraExtraLarge
+import platform.UIKit.UIContentSizeCategoryExtraExtraLarge
+import platform.UIKit.UIContentSizeCategoryExtraLarge
+import platform.UIKit.UIContentSizeCategoryExtraSmall
+import platform.UIKit.UIContentSizeCategoryLarge
+import platform.UIKit.UIContentSizeCategoryMedium
+import platform.UIKit.UIContentSizeCategorySmall
+import platform.UIKit.UIContentSizeCategoryUnspecified
+import platform.UIKit.UIView
+import platform.UIKit.UIViewController
+
+/**
+ * We need DensityProvider as a separate interface for lazy initialization.
+ */
+internal interface DensityProvider {
+    val density: Density
+    operator fun invoke() = density
+}
+
+internal class DensityProviderImpl(
+    private val uiViewControllerProvider: () -> UIViewController,
+    private val viewProvider: () -> UIView,
+) : DensityProvider {
+
+    val uiViewController get() = uiViewControllerProvider()
+
+    override val density: Density
+        get() {
+            val contentSizeCategory =
+                uiViewController.traitCollection.preferredContentSizeCategory
+                    ?: UIContentSizeCategoryUnspecified
+
+            return Density(
+                density = viewProvider().contentScaleFactor.toFloat(),
+                fontScale = uiContentSizeCategoryToFontScaleMap[contentSizeCategory] ?: 1.0f
+            )
+        }
+}
+
+private val uiContentSizeCategoryToFontScaleMap = mapOf(
+    UIContentSizeCategoryExtraSmall to 0.8f,
+    UIContentSizeCategorySmall to 0.85f,
+    UIContentSizeCategoryMedium to 0.9f,
+    UIContentSizeCategoryLarge to 1f, // default preference
+    UIContentSizeCategoryExtraLarge to 1.1f,
+    UIContentSizeCategoryExtraExtraLarge to 1.2f,
+    UIContentSizeCategoryExtraExtraExtraLarge to 1.3f,
+
+    // These values don't work well if they match scale shown by
+    // Text Size control hint, because iOS uses non-linear scaling
+    // calculated by UIFontMetrics, while Compose uses linear.
+    UIContentSizeCategoryAccessibilityMedium to 1.4f, // 160% native
+    UIContentSizeCategoryAccessibilityLarge to 1.5f, // 190% native
+    UIContentSizeCategoryAccessibilityExtraLarge to 1.6f, // 235% native
+    UIContentSizeCategoryAccessibilityExtraExtraLarge to 1.7f, // 275% native
+    UIContentSizeCategoryAccessibilityExtraExtraExtraLarge to 1.8f, // 310% native
+
+    // UIContentSizeCategoryUnspecified
+)
