+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/MetalRedrawer.uikit.kt
@@ -33,7 +33,6 @@ import platform.QuartzCore.*
 import platform.UIKit.UIApplicationDidEnterBackgroundNotification
 import platform.UIKit.UIApplicationWillEnterForegroundNotification
 import platform.darwin.*
-import kotlin.math.roundToInt
 import org.jetbrains.skia.Rect
 import platform.Foundation.NSLock
 import platform.Foundation.NSTimeInterval
@@ -191,6 +190,7 @@ internal class InflightCommandBuffers(
 internal class MetalRedrawer(
     private val metalLayer: CAMetalLayer,
     private val callbacks: MetalRedrawerCallbacks,
+    private val transparency: Boolean,
 ) {
     // Workaround for KN compiler bug
     // Type mismatch: inferred type is objcnames.protocols.MTLDeviceProtocol but platform.Metal.MTLDeviceProtocol was expected
@@ -236,7 +236,7 @@ internal class MetalRedrawer(
 
             // If active, make metalLayer transparent, opaque otherwise.
             // Rendering into opaque CAMetalLayer allows direct-to-screen optimization.
-            metalLayer.setOpaque(!value)
+            metalLayer.setOpaque(!value && !transparency)
             metalLayer.drawsAsynchronously = !value
         }
 
@@ -334,7 +334,7 @@ internal class MetalRedrawer(
                 width.toFloat(),
                 height.toFloat()
             )).also { canvas ->
-                canvas.clear(Color.WHITE)
+                canvas.clear(if (transparency) Color.TRANSPARENT else Color.WHITE)
                 callbacks.render(canvas, lastRenderTimestamp)
             }
